Upgrade to wcslib 4.8.4.

+++ astropy/wcs/src/wcslib/C/wcsutil.c
@@ -41,7 +41,7 @@
 void wcsutil_blank_fill(int n, char c[])
 
 {
-  size_t k;
+  int k;
 
   for (k = strlen(c); k < n; k++) {
     c[k] = ' ';
